Relocate and optimize Ghost Aurora feature
- Moved Ghost Aurora files (Engine, Layer, Shader) and tests to a dedicated `aurora` sub-package.
- Implemented BOLT ⚡ optimization in `GhostAuroraEngine.kt` with an early-exit loop for DESC-sorted behavior logs.
- Updated imports in `SeatingChartScreen.kt`.
- Added a local README.md explaining the Neural Metaphor and technical implementation.
- Updated the status of Ghost Aurora in `.jules/ghost_todo.md` to reflect its completion.
