+++ .jules/ghost_todo.md
@@ -3,6 +3,7 @@
 ## 🚧 Active Construction
 
 ## ✅ Completed Enhancements
+- [DONE] 2028-09-04: Ghost Aurora - Procedural background aurora reflecting classroom behavioral climate in `/labs/ghost/aurora/`
 - [DONE] 2028-09-03: Ghost Scape - Neural Spatial Audio pings in `/labs/ghost/scape/`
 - [DONE] 2028-09-02: Ghost Prism - Dynamic Individual Student Themes in `/labs/ghost/prism/`
 - [DONE] 2028-09-01: Ghost Coral - Social Reef & Collaborative Growth Visualization in `/labs/ghost/coral/`
